A bit of Extensions cleanup
